apps/edtech_home/requirements.txt
# Requirements for the project
# --extra-index-url https://download.pytorch.org/whl/cu121

# torch==2.3.0
# torchvision==0.18.0
# torchaudio==2.3.0

# # Other dependencies
# streamlit==1.36.0
# boto3==1.34.145


openai-whisper==20231117
fastapi==0.111.1
st_pages==1.0.0
pdfplumber==0.11.2
python-multipart==0.0.9
numpy==1.26.4